chore(web): reusable plus button
