+++ apps/web/src/design-system/index.ts
@@ -1,4 +1,6 @@
 export { Button } from './button/Button';
+export { PlusButton } from './button/PlusButton';
+export type { IButtonProps } from './button/Button';
 export type { Size } from './button/Button';
 export { Checkbox } from './checkbox/Checkbox';
 export { colors, shadows } from './config';
