+++ apps/web/src/pages/brand/tabs/LayoutsListPage.tsx
@@ -8,12 +8,11 @@ import { useState } from 'react';
 import { deleteLayoutById } from '../../../api/layouts';
 import { QueryKeys } from '../../../api/query.keys';
 import { When } from '../../../components/utils/When';
-import { colors, Text, Tooltip, withCellLoading } from '../../../design-system';
+import { colors, Text, Tooltip, PlusButton, withCellLoading } from '../../../design-system';
 import { Edit, Trash } from '../../../design-system/icons';
 import { IExtendedColumn, Table } from '../../../design-system/table/Table';
 import { useEnvController, useLayouts } from '../../../hooks';
 import { errorMessage, successMessage } from '../../../utils/notifications';
-import PageHeaderToolbar from '../../integrations/components/PageHeaderToolbar';
 import { DeleteConfirmModal } from '../../templates/components/DeleteConfirmModal';
 import { LayoutEditor } from './LayoutEditor';
 
@@ -164,7 +163,7 @@ export function LayoutsListPage({ handleLayoutAnalytics }: LayoutsListPageProps)
             marginTop: '10px',
           }}
         >
-          <PageHeaderToolbar
+          <PlusButton
             isGradient={!readonly}
             label="Add New Layout"
             disabled={readonly || isLoading || isLoadingDelete}
